use floating point width and height
fix issue #807

element offsetWidth and offsetHeight are rounded to integers,
which was causing off by one errors when further rounded to
calculate top and left values. This fix uses the raw floating
point width and height from getBoundingClientRect().

The jittering described in the issue is easily reproduced by
resizing the developer window to squash the goban, certain
sizes are problematic. I've not been able to reproduce the issue
with this fix in place.