Implement 'context-menu'
The 'context-menu' is apparently global to
webContents. We assume here this is actually the case.
Thus, some triage has to be done before
either to prevent it, or to advertise sabaki
which context we're talking about, and let
him perform the triage.

sabaki implements 'context-menu' as of openContextMenu. This
redirects either to the play menu or the edit menu
depending on mode. We've transferred the play menu from
PlayBar to sabaki

App puts a tiny phantom image in an anchor on top of layers, which
does not preventDefault, and thus promotes 'contextmenu'
to 'context-menu'. sabaki.openContextMenu is thus
the response to the Menu keyboard when focus is on the
body element.

Comment box also does not preventDefault (via stopPropagation), and
thus its context menu gets to sabaki.openContextMenu.