Allows users to edit influence boxes when in Estimate/Scoring mode.
When in scoring or estimation mode, we now track any vertexes that the user clicks
via the state.estimateOverrides property.  In our app render method, we then
override the area influence map with these clicks.  Clicks cycle through black/white/
empty.  A +1, -1 is used in the equations to reflect the fact that black/white/empty values
in the influence map are tracked via values of [-1,0,1].